requirements.txt
asgiref==3.2.7
attrs==19.3.0
cffi==1.14.0
constantly==15.1.0
cryptography==2.9.2
dateparser==0.7.4
Django==3.0.6
django-js-asset==1.2.2
django-mptt==0.11.0
incremental==17.5.0
more-itertools==8.3.0
ngrok==0.0.1
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.2
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
regex==2020.5.14
six==1.14.0
sqlparse==0.3.1
tzlocal==2.1
w3lib==1.21.0
wcwidth==0.1.9
djangorestframework>=3.11.0